Add files via upload
Controls: 
W to move up
S to move down
Click with mouse on restart if you win or lose
If ball gets stuck somehow before end screen please close the game and open again

This is a game, and really more of an art piece. The point of it is to show the advancement of the common ai accessible to anyone who has access to the internet.  This is a game of pong, its diffrent from normal pong, firstly because its very weird the ai you play against might as well be cheating and its very hard to beat. But the real reason why its diffrent is because the entire script for the game was written by an Ai (artificial intelligence) chatbot on the chat.openai website. The game was made in unity by me and 1 other friend. But the only thing we did was place the objects, write some text for the end screen, you win and lose screen. And make extremely minor adjustments to the game that would account for less than 5% of any actual coding done by a human being. All of the rest of the scripting/coding was done by asking an ai in the chat bot website a promt. Like "I want the game to do this" And the ai spitting out a script, which i then copied and pasted into unity. Please enjoy this awful game, and enjoy your future in our dystopian society ♥